refactor/Componenet folder structure rework
