+++ frontend/src/components/UI/Search/SearchBar.vue
