+++ frontend/src/components/UI/Search/SearchHeader.vue
